% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateClusterSimilarities.R
\name{calculateClusterSimilarities}
\alias{calculateClusterSimilarities}
\title{Function to find clusters in sparse matrix and then for each cluster row calculate distance of that row to the concept vector of cluster}
\usage{
calculateClusterSimilarities(sparseX, num_clusters)
}
\arguments{
\item{sparseX}{(sparce matrix with all data),num_clusters}
}
\value{
list with similarities from each row to its cluster center
}
\description{
Function to find clusters in sparse matrix and then for each cluster row calculate distance of that row to the concept vector of cluster
}
