% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineOptimalK.R
\name{determineOptimalK}
\alias{determineOptimalK}
\title{Function that evaluates an optimal nubmer of clusters per sample based on samples from both training data (randomly selected sample) and original data (divided into samples)}
\usage{
determineOptimalK(discretised_numeric_columns, mapped_symbolic_columns,
  discretisation, samples, samplesize, min_clusters, max_clusters)
}
\value{
a tuple 2 values: optimal k determined from training data, optimal k determined from original data
}
\description{
Function that evaluates an optimal nubmer of clusters per sample based on samples from both training data (randomly selected sample) and original data (divided into samples)
}
