% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAnomalousClusters.R
\name{findAnomalousClusters}
\alias{findAnomalousClusters}
\title{Function that performs training and application of the trained model on the data}
\usage{
findAnomalousClusters(discretised_numeric_columns, mapped_symbolic_columns,
  discretisation, samples, samplesize, optimal_k, data)
}
\value{
a list of anomalies sorted by sum of decision values
}
\description{
Function that performs training and application of the trained model on the data
}
