% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptimalK.R
\name{findOptimalK}
\alias{findOptimalK}
\title{Function that finds optimal k (number of clusters) for train samples using "elbow" method}
\usage{
findOptimalK(sims)
}
\arguments{
\item{c("num_clusters", "sim_measure")}{is a tuple containing 2 lists: (1) different number of clusters in the data and (2) corresponding values of similarity measure}
}
\value{
optimal_k
}
\description{
Function that finds optimal k (number of clusters) for train samples using "elbow" method
}
