% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAndNormalise.R
\name{ReadAndNormalise}
\alias{ReadAndNormalise}
\title{Function for reading and normalising mixed data (continuous numeric and symbolic)}
\usage{
ReadAndNormalise(data, column_types, discretisation)
}
\value{
a tuple with 3 matrices: discretised_numeric_columns, symbolic_columns_mapped_to_category, class_columns
}
\description{
Function for reading and normalising mixed data (continuous numeric and symbolic)
}
