% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorList2Matrix.R
\name{vectorList2Matrix}
\alias{vectorList2Matrix}
\title{Function to convert list of sparse vectors into sparse matrix.}
\usage{
vectorList2Matrix(vectorList)
}
\value{
sparse matrix
}
\description{
Function to convert list of sparse vectors into sparse matrix.
}
